Fix for kio.nread bug in SE.
SE appears to have a bug - occasionally kio.nread is erroneously set
to 0.  It looks like a memory management problem somewhere deep in
SE's code, since this problem is related to any memory allocation
calls elsewhere in the script.  For example, a call to "renew" inside
kstat traversing loop will cause nread to be 0 in the next iteration
of the loop.  "Data fixing" code, dealing with this issue, was removed
in revision 392.  This patch introduces a new function to re-read the
affected kstat, instead of ignoring bad data.  I can add a few more
"if" statements to cover the case when re-read data is still bad, but
I think this will be an overkill, since we should trust SE to some
degree :-).

* data_gatherers/orcallator/orcallator.se:
  (get_new_kstat_data):
    New function.
  (orca_io_info_update):
    Add code to re-read kstat data if kio.nread appears to be corrupt.